Modify main_page app test codes and views permission class
