--- templates/footer.html
@@ -1,62 +1,0 @@
-<!-- Footer -->
-<div class="container" id="globalfooter">
-	<footer class="footer" id="mg-global-footer">
-		<div class="conatainer">
-			{% if request.path != '/' %}
-			<section id="mg-gf-footer-logo">
-				<a href="/">
-					<img src="{{ static('Muserium_KO_EN.png') }}" alt="Brand Logo" class="img-fluid">
-				</a>
-			</section>
-			{% endif %}
-			<section class="mt-3" id="mg-gf-footer-legal">
-				<div class="row">
-					<div class="container d-flex justify-content-between">
-						<div class="footer-legal-copyright">
-							<p>Copyright © 2024 Muserium Glass. All rights reserved. Hosting by</p>
-						</div>
-						<div class="footer-legal-links d-flex">
-							<ul class="mg-gf-footer-legal-links list-inline">
-								<li class="mg-gf-footer-legal-link-item list-inline-item">
-									<router-link to="#" class="mg-gf-footer-legal-link">개인정보 처리방침</router-link>
-								</li>
-								<li class="list-inline-item">|</li>
-								<li class="mg-gf-footer-legal-link-item list-inline-item">
-									<router-link to="#" class="mg-gf-footer-legal-link">이용약관</router-link>
-								</li>
-								<li class="list-inline-item">|</li>
-								<li class="mg-gf-footer-legal-link-item list-inline-item">
-									<router-link to="#" class="mg-gf-footer-legal-link">이용안내</router-link>
-								</li>
-							</ul>
-						</div>
-					</div>
-				</div>
-			</section>
-			<section id="mg-gf-footer-info">
-				<div class="row align-items-center">
-					<div class="col-md">
-						<div class="footer-info">
-							<p id="brand-name">
-								<span class="footer-info-value text-black">{{ BRAND_NAME }} |</span>
-								<span class="footer-info-field">대표자명</span>
-								<span class="footer-info-value text-black">{{ CEO_NAME }} |</span>
-								<span class="footer-info-field text-black">주소</span>
-								<span class="footer-info-value text-black">{{ ADDRESS }} | </span>
-								<span class="footer-info-field text-black">전화</span>
-								<span class="footer-info-value text-black">{{ PHONE_NUMBER }} | </span>
-								<span class="footer-info-field text-black">사업자등록번호</span>
-								<span class="footer-info-value text-black">{{ BUSINESS_REGISTRATION_NUMBER }} | </span>
-								<span class="footer-info-field text-black">통신판매업신고번호</span>
-								<span class="footer-info-value text-black">{{ ECOMMERCE_REGISTRATION_NUMBER }}</span>
-								<router-link to="#" class="footer-info-value-link text-black">[사업자정보확인] | </router-link>
-								<span class="footer-info-field text-black">개인정보보호책임자</span>
-								<span class="footer-info-value text-black">{{ CPO_NAME }} | </span>
-							</p>
-						</div>
-					</div>
-				</div>
-			</section>
-		</div>
-	</footer>
-</div>
