--- templates/main_page/index.html
@@ -1,80 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<!-- Main banner Start -->
-<div class="container-fluid no-padding">
-    <div id="carouselExampleIndicators" class="carousel slide carousel-fade w-100 mx-auto bg-transparent" data-bs-ride="carousel">
-        {% if carousels %}
-            <div class="carousel-indicators">
-                {% for carousel in carousels %}
-                    <button type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide-to="{{ loop.index0 }}" class="{% if loop.first %}active{% endif %}" aria-current="{% if loop.first %}true{% endif %}" aria-label="{{ carousel.alt }}"></button>
-                {% endfor %}
-            </div>
-            <div class="carousel-inner carousel-height">
-                {% for carousel in carousels %}
-                    <div class="carousel-item {% if loop.first %}active{% endif %}">
-                        <img src="{{ static(carousel.image) }}" class="d-block mx-auto carousel-img-size img-fluid" alt="{{ carousel.alt }}" loading="lazy" onload="setBackgroundColor(this)">
-                    </div>
-                {% endfor %}
-            </div>
-        {% endif %}
-        <button class="carousel-control-prev" type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide="prev">
-            <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-            <span class="visually-hidden">이전</span>
-        </button>
-        <button class="carousel-control-next" type="button" data-bs-target="#carouselExampleIndicators" data-bs-slide="next">
-            <span class="carousel-control-next-icon" aria-hidden="true"></span>
-            <span class="visually-hidden">다음</span>
-        </button>
-    </div>
-</div>
-<!-- Main banner End -->
-<div>
-    <h1 class="text-center mt-5">Muserium Glass</h1>
-    <p class="text-center">유리가 있고 영감이 떠오르는 곳</p>
-</div>
-<!-- Card Area Start-->
-<div class="container mt-5">
-    <div class="row row-cols-1 row-cols-md-4 g-4">
-        {% if cards %}
-            {% for card in cards %}
-            <div class="col">
-                <div class="card h-100">
-                    <img src="{{ static(card.image) }}" class="card-img-top" alt="{{ card.alt }}" loading="lazy">
-                    <div class="card-body">
-                        <h5 class="card-title">{{ card.title }}</h5>
-                        <a href="{{ card.url }}" class="btn btn-dark">Learn More</a>
-                    </div>
-                </div>
-            </div>
-            {% endfor %}
-        {% endif %}
-    </div>
-</div>
-
-<!-- Card Area End -->
-<script>
-    // Set background color of carousel items
-    function setBackgroundColor(img) {
-        const colorThief = new ColorThief();
-        const color = colorThief.getColor(img);
-        img.closest('.carousel-item').style.backgroundColor = `rgb(${color[0]}, ${color[1]}, ${color[2]})`;
-    }
-    // Load color-thief script and set background color of carousel items
-    document.addEventListener('DOMContentLoaded', function() {
-        const script = document.createElement('script');
-        script.src = 'https://cdnjs.cloudflare.com/ajax/libs/color-thief/2.3.2/color-thief.umd.js';
-        script.onload = function() {
-            const images = document.querySelectorAll('.carousel-item img');
-            images.forEach(img => {
-                if (img.complete) {
-                    setBackgroundColor(img);
-                } else {
-                    img.addEventListener('load', () => setBackgroundColor(img));
-                }
-            });
-        };
-        document.head.appendChild(script);
-    });
-</script>
-{% endblock %}
