--- templates/navbar.html
@@ -1,5 +1,0 @@
-<!-- Navigation Bar -->
-<div id="app">
-	<navbar-component></navbar-component>
-</div>
-<script type="module" src="{{ static('src/main.js') }}"></script>
